handling some errors from login and session maintaining

+++ src/main/java/com/ecommerce/services/impls/LoginServicesImpl.java
@@ -32,7 +32,7 @@ public class LoginServicesImpl implements LoginServices {
 
     @Override
     public UserBean findUserById(int id) {
-        return UserMapper.INSTANCE.userEntityToUserBean(userRepository.findById((long) id));
+        return UserMapper.INSTANCE.userEntityToUserBean(userRepository.findById(id));
     }
